feat: bye bye BreezeTools
