+++ Sources/Breeze/Model/Alert.php
@@ -2,6 +2,8 @@
 
 declare(strict_types=1);
 
+namespace Breeze\Model;
+
 use Breeze\Breeze as Breeze;
 use Breeze\Entity\Alert as AlertEntity;
