+++ Sources/Breeze/Model/Like.php
@@ -2,6 +2,8 @@
 
 declare(strict_types=1);
 
+namespace Breeze\Model;
+
 use Breeze\Entity\Like as LikeEntity;
 
 class Like extends Base
