+++ Sources/Breeze/Model/Log.php
@@ -2,6 +2,7 @@
 
 declare(strict_types=1);
 
+namespace Breeze\Model;
 
 use Breeze\Entity\Log as LogEntity;
